LayoutTests/platform/mac-monterey/imported/w3c/web-platform-tests/fetch/private-network-access/mixed-content-fetch.tentative.https.window-expected.txt


FAIL https-local to http-local: wrong targetAddressSpace "unknown". assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-local to http-local: wrong targetAddressSpace "private". assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-local to http-local: wrong targetAddressSpace "public". assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-local to http-local: not a private network request. assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-local to http-private: wrong targetAddressSpace "unknown". assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-local to http-private: wrong targetAddressSpace "local". assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-local to http-private: wrong targetAddressSpace "public". assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-local to http-private: not a private network request. assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-local to http-public: wrong targetAddressSpace "unknown". assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-local to http-public: wrong targetAddressSpace "local". assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-local to http-public: wrong targetAddressSpace "private". assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-local to http-public: not a private network request. assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-private to http-local: missing targetAddressSpace. assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-private to http-local: wrong targetAddressSpace "unknown". assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-private to http-local: wrong targetAddressSpace "private". assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-private to http-local: wrong targetAddressSpace "public". assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-private to http-local: failed preflight. assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-private to http-local: success. assert_equals: error expected (undefined) undefined but got (string) "TypeError: Load failed"
FAIL https-private to http-local: PUT success. assert_equals: error expected (undefined) undefined but got (string) "TypeError: Load failed"
FAIL https-private to http-local: no-cors success. assert_equals: error expected (undefined) undefined but got (string) "TypeError: Load failed"
FAIL https-private to http-private: wrong targetAddressSpace "unknown". assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-private to http-private: wrong targetAddressSpace "local". assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-private to http-private: wrong targetAddressSpace "public". assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-private to http-private: not a private network request. assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-private to http-public: wrong targetAddressSpace "unknown". assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-private to http-public: wrong targetAddressSpace "local". assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-private to http-public: wrong targetAddressSpace "private". assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-private to http-public: not a private network request. assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-public to http-local: missing targetAddressSpace. assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-public to http-local: wrong targetAddressSpace "unknown". assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-public to http-local: wrong targetAddressSpace "private". assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-public to http-local: wrong targetAddressSpace "public". assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-public to http-local: failed preflight. assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-public to http-local: success. assert_equals: error expected (undefined) undefined but got (string) "TypeError: Load failed"
FAIL https-public to http-local: PUT success. assert_equals: error expected (undefined) undefined but got (string) "TypeError: Load failed"
FAIL https-public to http-local: no-cors success. assert_equals: error expected (undefined) undefined but got (string) "TypeError: Load failed"
FAIL https-public to http-private: missing targetAddressSpace. assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-public to http-private: wrong targetAddressSpace "unknown". assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-public to http-private: wrong targetAddressSpace "local". assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-public to http-private: wrong targetAddressSpace "public". assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-public to http-private: failed preflight. assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-public to http-private: success. assert_equals: error expected (undefined) undefined but got (string) "TypeError: Load failed"
FAIL https-public to http-private: PUT success. assert_equals: error expected (undefined) undefined but got (string) "TypeError: Load failed"
FAIL https-public to http-private: no-cors success. assert_equals: error expected (undefined) undefined but got (string) "TypeError: Load failed"
FAIL https-public to http-public: wrong targetAddressSpace "unknown". assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-public to http-public: wrong targetAddressSpace "local". assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-public to http-public: wrong targetAddressSpace "private". assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-public to http-public: not a private network request. assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-treat-as-public to http-local: wrong targetAddressSpace "private". assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-treat-as-public to http-local: success. assert_equals: error expected (undefined) undefined but got (string) "TypeError: Load failed"
FAIL https-treat-as-public to http-private: wrong targetAddressSpace "local". assert_equals: error expected "TypeError: Failed to fetch" but got "TypeError: Load failed"
FAIL https-treat-as-public to http-private: success. assert_equals: error expected (undefined) undefined but got (string) "TypeError: Load failed"
